requirements.txt
mysql-connector-python
uvicorn
fastapi[standard]
<<<<<<< HEAD
typing
pydantic
boto3
mysql-connector-python
chalice
pymongo
dnspython
mysqlclient
=======
mysql-connector-python

# typing
# pydantic
# pymongo
# dnspython
# mysqlclient
>>>>>>> faaf7b6123dda43b76098885671bb7a833f95988
